✨Introduce hot cache
Introduce the concept of component registration & deregistration of methodful operations specified in useResources. Only the first component to register for an operation will make ajax request for that operation (fetch data and the data is marked hot). Subsequent components to register won't make any ajax request but use the hot data. The last component to deregister will mark data as cold or stale with option to reset the operation state. The first component to register for an operation can choose to render with stale data for that operation while awaiting newly fetched data.

⚡️Enhancements to hook
  - Remove the need of StoreContext since we are using redux hooks useSelector/useDispatch
  - Remove buggy implementation of useDerivedState (calling setDerivedState in the render path is similar to calling this.setState in render method, and will trigger infinite loop of re-rendering)
  - Remove returning general status since it is difficult to calculate accurately & not that useful to the component